wip! Do two video passes and add zoom animation
This reworks things to process the transformation matrices
ahead of the output rendering so we can see ahead of time
if we should apply the matrix or not. it also adds a zoom animation.

It still needs to notice when part of the monitor is off camera so
the transform matrix doesn't abruptly zoom in and out as the aspect
ratio changes.

I guess we'll have to somehow extrapolate how much of the monitor
is off screen based on prior good frames to figure out monitor size.